feat: finish functional example

+++ pkg/protocol/protocol.go
@@ -114,7 +114,7 @@ type IndexRecord struct {
 	DataNumber uint64
 	// FieldByteOffset represents the byte offset of the field in the data
 	// file to fetch exactly the field value.
-	FieldStartByteOffset uint32
+	FieldStartByteOffset uint64
 	// FieldLength is pessimistic: it is an encoded value that is at least as
 	// long as the actual field value.
 	FieldLength int
